requirements.txt
langchain
langchain-openai
langchain-community
qdrant-client
python-docx
python-multipart
uvicorn
fastapi
python-dotenv
tiktoken
pymupdf
pydantic
pytest
pytest-asyncio
cohere
rank_bm25
langgraph>=0.2.0
docx2txt
